+++ email/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Flattened `/domain` folder.
+
 ## [0.11.0] - 2023-06-15
 
 ### Added